Added upload of audio file to Gcloud

+++ whisper_api/main.py
@@ -2,16 +2,19 @@
 Uses OpenAI's Whisper API to do speech-to-text transcription
 of audio, and Modal for easy containerized deployment of the app.
 """
+from . import audio, logger, video, gcloud, pdf, supabase, transcribe
+from .constants import ( CACHE_DIR, DEFAULT_MODEL, MODEL_DIR, RAW_AUDIO_DIR,
+                        TRANSCRIPTIONS_DIR )
 import json
 import os
 import pathlib
 from typing import Optional
 from fastapi import FastAPI
 from modal import Dict, Image, NetworkFileSystem, Stub, asgi_app, Secret
-from . import audio, logger, video, gcloud, pdf, supabase, transcribe
-from .constants import ( CACHE_DIR, DEFAULT_MODEL, MODEL_DIR, RAW_AUDIO_DIR,
-                        TRANSCRIPTIONS_DIR )
-
+import dacite
+import whisper
+import yt_dlp
+from google.oauth2 import service_account
 
 logger = logger.get_logger(name=__name__)
 
@@ -19,7 +22,7 @@ logger = logger.get_logger(name=__name__)
 volume = NetworkFileSystem.persisted(label="dataset-cache-vol")
 
 app_image = (
-    Image.debian_slim("3.11")
+    Image.debian_slim(python_version="3.11")
     .pip_install(
         "numba>=0.57.1",
         "fastapi",
@@ -74,23 +77,16 @@ def fastapi_app() -> FastAPI:
         Secret.from_name(app_name="supabase")
     ]
 )
-def process_audio(src_url: str, unique_id: int, session_title: Optional[str] = None, presenters: Optional[str] = None, is_video: bool=False, password: str=None) -> str:
-    import dacite
-    import whisper
-    import yt_dlp
-    from google.oauth2 import service_account
-    
+def process_audio(src_url: str, unique_id: int, session_title: Optional[str] = None, presenters: Optional[str] = None, is_video: bool=False, password: str=None) -> None:    
     # Get the title slug from the unique_id
     title_slug = str(object=unique_id)
 
-    destination_path = RAW_AUDIO_DIR / title_slug
-
-    # Video files are converted to mp3, so we need to pass the mp3 file path.
-    audio_filepath = f"{destination_path}.mp3" if is_video else destination_path
+    # Get working file paths from the slug
+    destination_path: pathlib.Path = RAW_AUDIO_DIR / title_slug
+    audio_filepath: pathlib.Path = f"{destination_path}.mp3" if is_video else destination_path
+    transcription_path: pathlib.Path = TRANSCRIPTIONS_DIR / f"{title_slug}.json"
 
     try:
-        transcription_path = get_transcript_path(title_slug=title_slug)
-
         # pre-download the model to the cache path, because the _download fn is not
         # thread-safe and can cause issues when multiple requests are made at the same time.
         model = DEFAULT_MODEL
@@ -101,13 +97,19 @@ def process_audio(src_url: str, unique_id: int, session_title: Optional[str] = N
 
         if is_video:
             video.download_convert_video_to_audio(
-                yt_dlp=yt_dlp, video_url=src_url, password=password, destination_path=destination_path
+                yt_dlp=yt_dlp,
+                video_url=src_url,
+                password=password,
+                destination_path=destination_path
             )
         else:
             audio.store_original_audio(
                 url=src_url,
                 destination=destination_path,
             )
+
+        # Upload the media file to Gcloud and get the public url
+        public_audio_url = gcloud.upload_to_gcloud(path=audio_filepath, credentials=credentials)
 
         logger.info(
             msg=f"Using the {model.name} model which has {model.params} parameters."
@@ -141,21 +143,33 @@ def process_audio(src_url: str, unique_id: int, session_title: Optional[str] = N
         credentials = service_account.Credentials.from_service_account_info(info=service_account_info)
 
         # Upload the PDF to Gcloud and get the public url
-        public_url = gcloud.upload_to_gcloud(pdf_path=pdf_path, credentials=credentials)
+        public_pdf_url = gcloud.upload_to_gcloud(path=pdf_path, credentials=credentials)
 
         # Upsert the transcript to Supabase
-        supabase.supabase_upsert(unique_id=unique_id, session_name=session_title if session_title else title_slug, presenters=presenters if presenters else "Unknown", transcript_text=output_text, audio_file_path=audio_filepath, transcript_file_path=public_url)
+        supabase.supabase_upsert(unique_id=unique_id, session_name=session_title if session_title else title_slug, presenters=presenters if presenters else "Unknown", transcript_text=output_text, audio_file_path=public_audio_url, transcript_file_path=public_pdf_url)
 
     except Exception as e:
         logger.exception(msg=e)
         raise dacite.DaciteError("Failed to process audio") from e
 
     finally:
-        logger.info(msg=f"Deleting the audio file in '{destination_path}'")
-        os.remove(path=audio_filepath)
-        logger.info(msg=f"Deleted the audio file in '{destination_path}'")
+        # Delete the audio file
+        if audio_filepath and os.path.exists(path=audio_filepath):
+            logger.info(msg=f"Deleting the audio file in '{audio_filepath}'")
+            os.remove(path=audio_filepath)
+            logger.info(msg=f"Deleted the audio file in '{audio_filepath}'")
 
-    return public_url  # return the public URL of the uploaded PDF
+        # Delete the transcription file
+        if transcription_path and os.path.exists(path=transcription_path):
+            logger.info(msg=f"Deleting the transcription file at '{transcription_path}'")
+            os.remove(path=transcription_path)
+            logger.info(msg=f"Deleted the transcription file at '{transcription_path}'")
+
+        # Delete the PDF file
+        if pdf_path and os.path.exists(path=pdf_path):
+            logger.info(msg=f"Deleting the PDF file at '{pdf_path}'")
+            os.remove(path=pdf_path)
+            logger.info(msg=f"Deleted the PDF file at '{pdf_path}'")
 
 
 # Register the transcribe_segment function with the stub
@@ -166,7 +180,3 @@ def process_audio(src_url: str, unique_id: int, session_title: Optional[str] = N
 )
 def transcribe_segment_wrapper(*args, **kwargs):
     return transcribe.transcribe_segment(*args, **kwargs)
-
-
-def get_transcript_path(title_slug: str) -> pathlib.Path:
-    return TRANSCRIPTIONS_DIR / f"{title_slug}.json"
